KAFKA-7119: Handle transient Kerberos errors on server side
